refactor: move game settings grom /mock/ folder to /src/
